Refactoring more tests.
