+++ cmd/fleetctl/fleetctl/gitops_test.go
@@ -2,8 +2,6 @@ package fleetctl
 
 import (
 	"context"
-	"database/sql"
-	"encoding/json"
 	"fmt"
 	"net/http"
 	"net/http/httptest"
@@ -11,11 +9,10 @@ import (
 	"path/filepath"
 	"slices"
 	"strings"
-	"sync"
 	"testing"
 	"time"
 
-	"github.com/fleetdm/fleet/v4/cmd/fleetctl/fleetctl/testingutils"
+	"github.com/fleetdm/fleet/v4/cmd/fleetctl/fleetctl/testing_utils"
 	"github.com/fleetdm/fleet/v4/pkg/file"
 	"github.com/fleetdm/fleet/v4/server/config"
 	"github.com/fleetdm/fleet/v4/server/datastore/mysql"
@@ -23,14 +20,12 @@ import (
 	apple_mdm "github.com/fleetdm/fleet/v4/server/mdm/apple"
 	"github.com/fleetdm/fleet/v4/server/mdm/apple/vpp"
 	"github.com/fleetdm/fleet/v4/server/mdm/nanodep/tokenpki"
-	"github.com/fleetdm/fleet/v4/server/mdm/testing_utils"
-	"github.com/fleetdm/fleet/v4/server/mock"
+	mdmtesting "github.com/fleetdm/fleet/v4/server/mdm/testing_utils"
 	digicert_mock "github.com/fleetdm/fleet/v4/server/mock/digicert"
 	mdmmock "github.com/fleetdm/fleet/v4/server/mock/mdm"
 	scep_mock "github.com/fleetdm/fleet/v4/server/mock/scep"
 	"github.com/fleetdm/fleet/v4/server/ptr"
 	"github.com/fleetdm/fleet/v4/server/service"
-	"github.com/fleetdm/fleet/v4/server/test"
 	"github.com/google/uuid"
 	"github.com/jmoiron/sqlx"
 	"github.com/stretchr/testify/assert"
@@ -43,46 +38,6 @@ const (
 	orgName        = "GitOps Test"
 )
 
-func addLabelMocks(ds *mock.Store) {
-	var deletedLabels []string
-	ds.GetLabelSpecsFunc = func(ctx context.Context) ([]*fleet.LabelSpec, error) {
-		return []*fleet.LabelSpec{
-			{
-				Name:                "a",
-				Description:         "A global label",
-				LabelMembershipType: fleet.LabelMembershipTypeManual,
-				Hosts:               []string{"host2", "host3"},
-			},
-			{
-				Name:                "b",
-				Description:         "Another global label",
-				LabelMembershipType: fleet.LabelMembershipTypeDynamic,
-				Query:               "SELECT 1 from osquery_info",
-			},
-		}, nil
-	}
-	ds.ApplyLabelSpecsWithAuthorFunc = func(ctx context.Context, specs []*fleet.LabelSpec, authorID *uint) (err error) {
-		return nil
-	}
-
-	ds.DeleteLabelFunc = func(ctx context.Context, name string) error {
-		deletedLabels = append(deletedLabels, name)
-		return nil
-	}
-	ds.LabelsByNameFunc = func(ctx context.Context, names []string) (map[string]*fleet.Label, error) {
-		return map[string]*fleet.Label{
-			"a": {
-				ID:   1,
-				Name: "a",
-			},
-			"b": {
-				ID:   2,
-				Name: "b",
-			},
-		}, nil
-	}
-}
-
 func TestGitOpsFilenameValidation(t *testing.T) {
 	filename := strings.Repeat("a", filenameMaxLength+1)
 	_, err := RunAppNoChecks([]string{"gitops", "-f", filename})
@@ -92,7 +47,7 @@ func TestGitOpsFilenameValidation(t *testing.T) {
 func TestGitOpsBasicGlobalFree(t *testing.T) {
 	// Cannot run t.Parallel() because it sets environment variables
 
-	_, ds := testingutils.RunServerWithMockedDS(t)
+	_, ds := testing_utils.RunServerWithMockedDS(t)
 
 	ds.BatchSetMDMProfilesFunc = func(
 		ctx context.Context, tmID *uint, macProfiles []*fleet.MDMAppleConfigProfile, winProfiles []*fleet.MDMWindowsConfigProfile,
@@ -258,10 +213,10 @@ func TestGitOpsBasicGlobalPremium(t *testing.T) {
 	scepConfig.ValidateNDESSCEPAdminURLFunc = func(_ context.Context, _ fleet.NDESSCEPProxyIntegration) error { return nil }
 	digiCertService := &digicert_mock.Service{}
 	digiCertService.VerifyProfileIDFunc = func(_ context.Context, _ fleet.DigiCertIntegration) error { return nil }
-	_, ds := testingutils.RunServerWithMockedDS(
+	_, ds := testing_utils.RunServerWithMockedDS(
 		t, &service.TestServerOpts{
 			License:           license,
-			KeyValueStore:     newMemKeyValueStore(),
+			KeyValueStore: testing_utils.NewMemKeyValueStore(),
 			EnableSCEPProxy:   true,
 			SCEPConfigService: scepConfig,
 			DigiCertService:   digiCertService,
@@ -483,10 +438,10 @@ software:
 func TestGitOpsBasicTeam(t *testing.T) {
 	// Cannot run t.Parallel() because it sets environment variables
 	license := &fleet.LicenseInfo{Tier: fleet.TierPremium, Expiration: time.Now().Add(24 * time.Hour)}
-	_, ds := testingutils.RunServerWithMockedDS(
+	_, ds := testing_utils.RunServerWithMockedDS(
 		t, &service.TestServerOpts{
 			License:       license,
-			KeyValueStore: newMemKeyValueStore(),
+			KeyValueStore: testing_utils.NewMemKeyValueStore(),
 		},
 	)
 
@@ -690,10 +645,10 @@ func TestGitOpsFullGlobal(t *testing.T) {
 	config.SetTestMDMConfig(t, &fleetCfg, testCertPEM, testKeyPEM, "../../../server/service/testdata")
 
 	// License is not needed because we are not using any premium features in our config.
-	_, ds := testingutils.RunServerWithMockedDS(
+	_, ds := testing_utils.RunServerWithMockedDS(
 		t, &service.TestServerOpts{
 			MDMStorage:  new(mdmmock.MDMAppleStore),
-			MDMPusher:   mockPusher{},
+			MDMPusher: testing_utils.MockPusher{},
 			FleetConfig: &fleetCfg,
 		},
 	)
@@ -963,14 +918,14 @@ func TestGitOpsFullTeam(t *testing.T) {
 	config.SetTestMDMConfig(t, &fleetCfg, testCertPEM, testKeyPEM, "../../../server/service/testdata")
 
 	// License is not needed because we are not using any premium features in our config.
-	_, ds := testingutils.RunServerWithMockedDS(
+	_, ds := testing_utils.RunServerWithMockedDS(
 		t, &service.TestServerOpts{
 			License:          license,
 			MDMStorage:       new(mdmmock.MDMAppleStore),
-			MDMPusher:        mockPusher{},
+			MDMPusher:     testing_utils.MockPusher{},
 			FleetConfig:      &fleetCfg,
 			NoCacheDatastore: true,
-			KeyValueStore:    newMemKeyValueStore(),
+			KeyValueStore: testing_utils.NewMemKeyValueStore(),
 		},
 	)
 
@@ -1166,7 +1121,7 @@ func TestGitOpsFullTeam(t *testing.T) {
 		return nil
 	}
 
-	addLabelMocks(ds)
+	testing_utils.AddLabelMocks(ds)
 
 	var appliedSoftwareInstallers []*fleet.UploadSoftwareInstallerPayload
 	ds.BatchSetSoftwareInstallersFunc = func(ctx context.Context, teamID *uint, installers []*fleet.UploadSoftwareInstallerPayload) error {
@@ -1195,7 +1150,7 @@ func TestGitOpsFullTeam(t *testing.T) {
 		return nil
 	}
 
-	testingutils.StartSoftwareInstallerServer(t)
+	testing_utils.StartSoftwareInstallerServer(t)
 
 	t.Setenv("TEST_TEAM_NAME", teamName)
 
@@ -1306,7 +1261,7 @@ software:
 }
 
 func createFakeITunesAndVPPServices(t *testing.T) {
-	config := &appleVPPConfigSrvConf{
+	config := &testing_utils.AppleVPPConfigSrvConf{
 		Assets: []vpp.Asset{
 			{
 				AdamID:         "1",
@@ -1321,7 +1276,7 @@ func createFakeITunesAndVPPServices(t *testing.T) {
 		},
 		SerialNumbers: []string{"123", "456"},
 	}
-	startVPPApplyServer(t, config)
+	testing_utils.StartVPPApplyServer(t, config)
 
 	appleITunesSrv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		// a map of apps we can respond with
@@ -1352,10 +1307,10 @@ func createFakeITunesAndVPPServices(t *testing.T) {
 func TestGitOpsBasicGlobalAndTeam(t *testing.T) {
 	// Cannot run t.Parallel() because it sets environment variables
 	license := &fleet.LicenseInfo{Tier: fleet.TierPremium, Expiration: time.Now().Add(24 * time.Hour)}
-	_, ds := testingutils.RunServerWithMockedDS(
+	_, ds := testing_utils.RunServerWithMockedDS(
 		t, &service.TestServerOpts{
 			License:       license,
-			KeyValueStore: newMemKeyValueStore(),
+			KeyValueStore: testing_utils.NewMemKeyValueStore(),
 		},
 	)
 
@@ -1446,7 +1401,7 @@ func TestGitOpsBasicGlobalAndTeam(t *testing.T) {
 		return nil, 0, nil, nil
 	}
 
-	addLabelMocks(ds)
+	testing_utils.AddLabelMocks(ds)
 
 	ds.NewActivityFunc = func(
 		ctx context.Context, user *fleet.User, activity fleet.ActivityDetails, details []byte, createdAt time.Time,
@@ -1704,10 +1659,10 @@ func TestGitOpsBasicGlobalAndNoTeam(t *testing.T) {
 	// environment variable.
 
 	license := &fleet.LicenseInfo{Tier: fleet.TierPremium, Expiration: time.Now().Add(24 * time.Hour)}
-	_, ds := testingutils.RunServerWithMockedDS(
+	_, ds := testing_utils.RunServerWithMockedDS(
 		t, &service.TestServerOpts{
 			License:       license,
-			KeyValueStore: newMemKeyValueStore(),
+			KeyValueStore: testing_utils.NewMemKeyValueStore(),
 		},
 	)
 	// Mock appConfig
@@ -1788,7 +1743,7 @@ func TestGitOpsBasicGlobalAndNoTeam(t *testing.T) {
 	ds.ListQueriesFunc = func(ctx context.Context, opts fleet.ListQueryOptions) ([]*fleet.Query, int, *fleet.PaginationMetadata, error) {
 		return nil, 0, nil, nil
 	}
-	addLabelMocks(ds)
+	testing_utils.AddLabelMocks(ds)
 
 	ds.NewActivityFunc = func(
 		ctx context.Context, user *fleet.User, activity fleet.ActivityDetails, details []byte, createdAt time.Time,
@@ -2149,8 +2104,8 @@ software:
 func TestGitOpsFullGlobalAndTeam(t *testing.T) {
 	// Cannot run t.Parallel() because it sets environment variables
 	// mdm test configuration must be set so that activating windows MDM works.
-	ds, savedAppConfigPtr, savedTeams := setupFullGitOpsPremiumServer(t)
-	testingutils.StartSoftwareInstallerServer(t)
+	ds, savedAppConfigPtr, savedTeams := testing_utils.SetupFullGitOpsPremiumServer(t)
+	testing_utils.StartSoftwareInstallerServer(t)
 
 	var enrolledSecrets []*fleet.EnrollSecret
 	var enrolledTeamSecrets []*fleet.EnrollSecret
@@ -2203,7 +2158,7 @@ func TestGitOpsFullGlobalAndTeam(t *testing.T) {
 		return []uint{}, nil
 	}
 
-	apnsCert, apnsKey, err := mysql.GenerateTestCertBytes(testing_utils.NewTestMDMAppleCertTemplate())
+	apnsCert, apnsKey, err := mysql.GenerateTestCertBytes(mdmtesting.NewTestMDMAppleCertTemplate())
 	require.NoError(t, err)
 	crt, key, err := apple_mdm.NewSCEPCACertKey()
 	require.NoError(t, err)
@@ -2343,532 +2298,6 @@ software:
 	})
 }
 
-func TestGitOpsTeamSofwareInstallers(t *testing.T) {
-	testingutils.StartSoftwareInstallerServer(t)
-	startAndServeVPPServer(t)
-
-	cases := []struct {
-		file    string
-		wantErr string
-	}{
-		{"testdata/gitops/team_software_installer_not_found.yml", "Please make sure that URLs are reachable from your Fleet server."},
-		{"testdata/gitops/team_software_installer_unsupported.yml", "The file should be .pkg, .msi, .exe, .deb, .rpm, or .tar.gz."},
-		// commenting out, results in the process getting killed on CI and on some machines
-		// {"testdata/gitops/team_software_installer_too_large.yml", "The maximum file size is 3 GB"},
-		{"testdata/gitops/team_software_installer_valid.yml", ""},
-		{"testdata/gitops/team_software_installer_subdir.yml", ""},
-		{"testdata/gitops/subdir/team_software_installer_valid.yml", ""},
-		{"testdata/gitops/team_software_installer_valid_apply.yml", ""},
-		{"testdata/gitops/team_software_installer_pre_condition_multiple_queries.yml", "should have only one query."},
-		{"testdata/gitops/team_software_installer_pre_condition_multiple_queries_apply.yml", "should have only one query."},
-		{"testdata/gitops/team_software_installer_pre_condition_not_found.yml", "no such file or directory"},
-		{"testdata/gitops/team_software_installer_install_not_found.yml", "no such file or directory"},
-		{"testdata/gitops/team_software_installer_uninstall_not_found.yml", "no such file or directory"},
-		{"testdata/gitops/team_software_installer_post_install_not_found.yml", "no such file or directory"},
-		{"testdata/gitops/team_software_installer_no_url.yml", "at least one of hash_sha256 or url is required for each software package"},
-		{"testdata/gitops/team_software_installer_invalid_self_service_value.yml", "\"packages.SoftwarePackageSpec.self_service\" must be a bool, found string"},
-		{"testdata/gitops/team_software_installer_invalid_both_include_exclude.yml", `only one of "labels_exclude_any" or "labels_include_any" can be specified`},
-		{"testdata/gitops/team_software_installer_valid_include.yml", ""},
-		{"testdata/gitops/team_software_installer_valid_exclude.yml", ""},
-		{"testdata/gitops/team_software_installer_invalid_unknown_label.yml", "Please create the missing labels, or update your settings to not refer to these labels."},
-		// team tests for setup experience software/script
-		{"testdata/gitops/team_setup_software_valid.yml", ""},
-		{"testdata/gitops/team_setup_software_invalid_script.yml", "no_such_script.sh: no such file"},
-		{"testdata/gitops/team_setup_software_invalid_software_package.yml", "no_such_software.yml\" does not exist for that team"},
-		{"testdata/gitops/team_setup_software_invalid_vpp_app.yml", "\"no_such_app\" does not exist for that team"},
-	}
-	for _, c := range cases {
-		t.Run(filepath.Base(c.file), func(t *testing.T) {
-			ds, _, _ := setupFullGitOpsPremiumServer(t)
-			tokExpire := time.Now().Add(time.Hour)
-			token, err := test.CreateVPPTokenEncoded(tokExpire, "fleet", "ca")
-			require.NoError(t, err)
-
-			ds.SetTeamVPPAppsFunc = func(ctx context.Context, teamID *uint, adamIDs []fleet.VPPAppTeam) error {
-				return nil
-			}
-			ds.GetVPPAppsFunc = func(ctx context.Context, teamID *uint) ([]fleet.VPPAppResponse, error) {
-				return []fleet.VPPAppResponse{}, nil
-			}
-			ds.BatchInsertVPPAppsFunc = func(ctx context.Context, apps []*fleet.VPPApp) error {
-				return nil
-			}
-			ds.GetVPPTokenByTeamIDFunc = func(ctx context.Context, teamID *uint) (*fleet.VPPTokenDB, error) {
-				return &fleet.VPPTokenDB{
-					ID:        1,
-					OrgName:   "Fleet",
-					Location:  "Earth",
-					RenewDate: tokExpire,
-					Token:     string(token),
-					Teams:     nil,
-				}, nil
-			}
-
-			ds.GetLabelSpecsFunc = func(ctx context.Context) ([]*fleet.LabelSpec, error) {
-				return []*fleet.LabelSpec{
-					{
-						Name:                "a",
-						Description:         "A global label",
-						LabelMembershipType: fleet.LabelMembershipTypeManual,
-						Hosts:               []string{"host2", "host3"},
-					},
-					{
-						Name:                "b",
-						Description:         "Another label",
-						LabelMembershipType: fleet.LabelMembershipTypeDynamic,
-						Query:               "SELECT 1 from osquery_info",
-					},
-				}, nil
-			}
-
-			labelToIDs := map[string]uint{
-				fleet.BuiltinLabelMacOS14Plus: 1,
-				"a":                           2,
-				"b":                           3,
-			}
-			ds.LabelIDsByNameFunc = func(ctx context.Context, labels []string) (map[string]uint, error) {
-				// for this test, recognize labels a and b (as well as the built-in macos 14+ one)
-				ret := make(map[string]uint)
-				for _, lbl := range labels {
-					id, ok := labelToIDs[lbl]
-					if ok {
-						ret[lbl] = id
-					}
-				}
-				return ret, nil
-			}
-			ds.GetTeamsWithInstallerByHashFunc = func(ctx context.Context, sha256, url string) (map[uint]*fleet.ExistingSoftwareInstaller, error) {
-				return map[uint]*fleet.ExistingSoftwareInstaller{}, nil
-			}
-			ds.GetSoftwareCategoryIDsFunc = func(ctx context.Context, names []string) ([]uint, error) {
-				return []uint{}, nil
-			}
-
-			_, err = RunAppNoChecks([]string{"gitops", "-f", c.file})
-			if c.wantErr == "" {
-				require.NoError(t, err)
-			} else {
-				require.ErrorContains(t, err, c.wantErr)
-			}
-		})
-	}
-}
-
-func TestGitOpsTeamSoftwareInstallersQueryEnv(t *testing.T) {
-	testingutils.StartSoftwareInstallerServer(t)
-	ds, _, _ := setupFullGitOpsPremiumServer(t)
-
-	t.Setenv("QUERY_VAR", "IT_WORKS")
-
-	ds.BatchSetSoftwareInstallersFunc = func(ctx context.Context, tmID *uint, installers []*fleet.UploadSoftwareInstallerPayload) error {
-		if len(installers) != 0 && installers[0].PreInstallQuery != "select IT_WORKS" {
-			return fmt.Errorf("Missing env var, got %s", installers[0].PreInstallQuery)
-		}
-		return nil
-	}
-	ds.GetSoftwareInstallersFunc = func(ctx context.Context, tmID uint) ([]fleet.SoftwarePackageResponse, error) {
-		return nil, nil
-	}
-	ds.GetTeamsWithInstallerByHashFunc = func(ctx context.Context, sha256, url string) (map[uint]*fleet.ExistingSoftwareInstaller, error) {
-		return map[uint]*fleet.ExistingSoftwareInstaller{}, nil
-	}
-	ds.GetSoftwareCategoryIDsFunc = func(ctx context.Context, names []string) ([]uint, error) {
-		return []uint{}, nil
-	}
-
-	_, err := RunAppNoChecks([]string{"gitops", "-f", "testdata/gitops/team_software_installer_valid_env_query.yml"})
-	require.NoError(t, err)
-}
-
-func TestGitOpsNoTeamVPPPolicies(t *testing.T) {
-	startAndServeVPPServer(t)
-
-	cases := []struct {
-		noTeamFile string
-		wantErr    string
-		vppApps    []fleet.VPPAppResponse
-	}{
-		{
-			noTeamFile: "testdata/gitops/subdir/no_team_vpp_policies_valid.yml",
-			vppApps: []fleet.VPPAppResponse{
-				{ // for more test coverage
-					Platform: fleet.MacOSPlatform,
-				},
-				{ // for more test coverage
-					TitleID:  ptr.Uint(122),
-					Platform: fleet.MacOSPlatform,
-				},
-				{
-					TeamID:     ptr.Uint(0),
-					TitleID:    ptr.Uint(123),
-					AppStoreID: "1",
-					Platform:   fleet.IOSPlatform,
-				},
-				{
-					TeamID:     ptr.Uint(0),
-					TitleID:    ptr.Uint(124),
-					AppStoreID: "1",
-					Platform:   fleet.MacOSPlatform,
-				},
-				{
-					TeamID:     ptr.Uint(0),
-					TitleID:    ptr.Uint(125),
-					AppStoreID: "1",
-					Platform:   fleet.IPadOSPlatform,
-				},
-			},
-		},
-	}
-	for _, c := range cases {
-		t.Run(filepath.Base(c.noTeamFile), func(t *testing.T) {
-			ds, _, _ := setupFullGitOpsPremiumServer(t)
-			tokExpire := time.Now().Add(time.Hour)
-			token, err := test.CreateVPPTokenEncoded(tokExpire, "fleet", "ca")
-			require.NoError(t, err)
-
-			ds.SetTeamVPPAppsFunc = func(ctx context.Context, teamID *uint, adamIDs []fleet.VPPAppTeam) error {
-				return nil
-			}
-			ds.BatchInsertVPPAppsFunc = func(ctx context.Context, apps []*fleet.VPPApp) error {
-				return nil
-			}
-			ds.GetVPPAppsFunc = func(ctx context.Context, teamID *uint) ([]fleet.VPPAppResponse, error) {
-				return c.vppApps, nil
-			}
-			ds.GetVPPTokenByTeamIDFunc = func(ctx context.Context, teamID *uint) (*fleet.VPPTokenDB, error) {
-				return &fleet.VPPTokenDB{
-					ID:        1,
-					OrgName:   "Fleet",
-					Location:  "Earth",
-					RenewDate: tokExpire,
-					Token:     string(token),
-					Teams:     nil,
-				}, nil
-			}
-			labelToIDs := map[string]uint{
-				fleet.BuiltinLabelMacOS14Plus: 1,
-				"a":                           2,
-				"b":                           3,
-			}
-			ds.LabelIDsByNameFunc = func(ctx context.Context, labels []string) (map[string]uint, error) {
-				// for this test, recognize labels a and b (as well as the built-in macos 14+ one)
-				ret := make(map[string]uint)
-				for _, lbl := range labels {
-					id, ok := labelToIDs[lbl]
-					if ok {
-						ret[lbl] = id
-					}
-				}
-				return ret, nil
-			}
-			ds.LabelsByNameFunc = func(ctx context.Context, names []string) (map[string]*fleet.Label, error) {
-				return map[string]*fleet.Label{
-					"a": {
-						ID:   1,
-						Name: "a",
-					},
-					"b": {
-						ID:   2,
-						Name: "b",
-					},
-				}, nil
-			}
-			ds.GetSoftwareCategoryIDsFunc = func(ctx context.Context, names []string) ([]uint, error) {
-				return []uint{}, nil
-			}
-
-			t.Setenv("APPLE_BM_DEFAULT_TEAM", "")
-			globalFile := "./testdata/gitops/global_config_no_paths.yml"
-			dstPath := filepath.Join(filepath.Dir(c.noTeamFile), "no-team.yml")
-			t.Cleanup(func() {
-				os.Remove(dstPath)
-			})
-			err = file.Copy(c.noTeamFile, dstPath, 0o755)
-			require.NoError(t, err)
-			_, err = RunAppNoChecks([]string{"gitops", "-f", globalFile, "-f", dstPath})
-			if c.wantErr == "" {
-				require.NoError(t, err)
-			} else {
-				require.ErrorContains(t, err, c.wantErr)
-			}
-		})
-	}
-}
-
-func TestGitOpsNoTeamSoftwareInstallers(t *testing.T) {
-	testingutils.StartSoftwareInstallerServer(t)
-	startAndServeVPPServer(t)
-
-	cases := []struct {
-		noTeamFile string
-		wantErr    string
-	}{
-		{"testdata/gitops/no_team_software_installer_not_found.yml", "Please make sure that URLs are reachable from your Fleet server."},
-		{"testdata/gitops/no_team_software_installer_unsupported.yml", "The file should be .pkg, .msi, .exe, .deb, .rpm, or .tar.gz."},
-		// commenting out, results in the process getting killed on CI and on some machines
-		// {"testdata/gitops/no_team_software_installer_too_large.yml", "The maximum file size is 3 GB"},
-		{"testdata/gitops/no_team_software_installer_valid.yml", ""},
-		{"testdata/gitops/no_team_software_installer_subdir.yml", ""},
-		{"testdata/gitops/subdir/no_team_software_installer_valid.yml", ""},
-		{"testdata/gitops/no_team_software_installer_pre_condition_multiple_queries.yml", "should have only one query."},
-		{"testdata/gitops/no_team_software_installer_pre_condition_not_found.yml", "no such file or directory"},
-		{"testdata/gitops/no_team_software_installer_install_not_found.yml", "no such file or directory"},
-		{"testdata/gitops/no_team_software_installer_uninstall_not_found.yml", "no such file or directory"},
-		{"testdata/gitops/no_team_software_installer_post_install_not_found.yml", "no such file or directory"},
-		{"testdata/gitops/no_team_software_installer_no_url.yml", "at least one of hash_sha256 or url is required for each software package"},
-		{"testdata/gitops/no_team_software_installer_invalid_self_service_value.yml", "\"packages.SoftwarePackageSpec.self_service\" must be a bool, found string"},
-		{"testdata/gitops/no_team_software_installer_invalid_both_include_exclude.yml", `only one of "labels_exclude_any" or "labels_include_any" can be specified`},
-		{"testdata/gitops/no_team_software_installer_valid_include.yml", ""},
-		{"testdata/gitops/no_team_software_installer_valid_exclude.yml", ""},
-		{"testdata/gitops/no_team_software_installer_invalid_unknown_label.yml", "Please create the missing labels, or update your settings to not refer to these labels."},
-		// No team tests for setup experience software/script
-		{"testdata/gitops/no_team_setup_software_valid.yml", ""},
-		{"testdata/gitops/no_team_setup_software_invalid_script.yml", "no_such_script.sh: no such file"},
-		{"testdata/gitops/no_team_setup_software_invalid_software_package.yml", "no_such_software.yml\" does not exist for that team"},
-		{"testdata/gitops/no_team_setup_software_invalid_vpp_app.yml", "\"no_such_app\" does not exist for that team"},
-	}
-	for _, c := range cases {
-		t.Run(filepath.Base(c.noTeamFile), func(t *testing.T) {
-			ds, _, _ := setupFullGitOpsPremiumServer(t)
-			tokExpire := time.Now().Add(time.Hour)
-			token, err := test.CreateVPPTokenEncoded(tokExpire, "fleet", "ca")
-			require.NoError(t, err)
-
-			ds.SetTeamVPPAppsFunc = func(ctx context.Context, teamID *uint, adamIDs []fleet.VPPAppTeam) error {
-				return nil
-			}
-			ds.BatchInsertVPPAppsFunc = func(ctx context.Context, apps []*fleet.VPPApp) error {
-				return nil
-			}
-			ds.GetVPPAppsFunc = func(ctx context.Context, teamID *uint) ([]fleet.VPPAppResponse, error) {
-				return []fleet.VPPAppResponse{}, nil
-			}
-			ds.GetVPPTokenByTeamIDFunc = func(ctx context.Context, teamID *uint) (*fleet.VPPTokenDB, error) {
-				return &fleet.VPPTokenDB{
-					ID:        1,
-					OrgName:   "Fleet",
-					Location:  "Earth",
-					RenewDate: tokExpire,
-					Token:     string(token),
-					Teams:     nil,
-				}, nil
-			}
-			ds.GetLabelSpecsFunc = func(ctx context.Context) ([]*fleet.LabelSpec, error) {
-				return []*fleet.LabelSpec{
-					{
-						Name:                "a",
-						Description:         "A global label",
-						LabelMembershipType: fleet.LabelMembershipTypeManual,
-						Hosts:               []string{"host2", "host3"},
-					},
-					{
-						Name:                "b",
-						Description:         "Another label",
-						LabelMembershipType: fleet.LabelMembershipTypeDynamic,
-						Query:               "SELECT 1 from osquery_info",
-					},
-				}, nil
-			}
-			labelToIDs := map[string]uint{
-				fleet.BuiltinLabelMacOS14Plus: 1,
-				"a":                           2,
-				"b":                           3,
-			}
-			ds.LabelIDsByNameFunc = func(ctx context.Context, labels []string) (map[string]uint, error) {
-				// for this test, recognize labels a and b (as well as the built-in macos 14+ one)
-				ret := make(map[string]uint)
-				for _, lbl := range labels {
-					id, ok := labelToIDs[lbl]
-					if ok {
-						ret[lbl] = id
-					}
-				}
-				return ret, nil
-			}
-			ds.GetTeamsWithInstallerByHashFunc = func(ctx context.Context, sha256, url string) (map[uint]*fleet.ExistingSoftwareInstaller, error) {
-				return map[uint]*fleet.ExistingSoftwareInstaller{}, nil
-			}
-			ds.GetSoftwareCategoryIDsFunc = func(ctx context.Context, names []string) ([]uint, error) {
-				return []uint{}, nil
-			}
-
-			t.Setenv("APPLE_BM_DEFAULT_TEAM", "")
-			globalFile := "./testdata/gitops/global_config_no_paths.yml"
-			if strings.HasPrefix(filepath.Base(c.noTeamFile), "no_team_setup_software") {
-				// the controls section is in the no-team test file, so use a global file without that section
-				globalFile = "./testdata/gitops/global_config_no_paths_no_controls.yml"
-			}
-			dstPath := filepath.Join(filepath.Dir(c.noTeamFile), "no-team.yml")
-			t.Cleanup(func() {
-				os.Remove(dstPath)
-			})
-			err = file.Copy(c.noTeamFile, dstPath, 0o755)
-			require.NoError(t, err)
-			_, err = RunAppNoChecks([]string{"gitops", "-f", globalFile, "-f", dstPath})
-			if c.wantErr == "" {
-				require.NoError(t, err)
-			} else {
-				require.ErrorContains(t, err, c.wantErr)
-			}
-		})
-	}
-}
-
-func TestGitOpsTeamVPPApps(t *testing.T) {
-	startAndServeVPPServer(t)
-
-	cases := []struct {
-		file            string
-		wantErr         string
-		tokenExpiration time.Time
-		expectedLabels  map[string]uint
-	}{
-		{"testdata/gitops/team_vpp_valid_app.yml", "", time.Now().Add(24 * time.Hour), map[string]uint{}},
-		{"testdata/gitops/team_vpp_valid_app_self_service.yml", "", time.Now().Add(24 * time.Hour), map[string]uint{}},
-		{"testdata/gitops/team_vpp_valid_empty.yml", "", time.Now().Add(24 * time.Hour), map[string]uint{}},
-		{"testdata/gitops/team_vpp_valid_empty.yml", "", time.Now().Add(-24 * time.Hour), map[string]uint{}},
-		{"testdata/gitops/team_vpp_valid_app.yml", "VPP token expired", time.Now().Add(-24 * time.Hour), map[string]uint{}},
-		{"testdata/gitops/team_vpp_invalid_app.yml", "app not available on vpp account", time.Now().Add(24 * time.Hour), map[string]uint{}},
-		{"testdata/gitops/team_vpp_incorrect_type.yml", "\"app_store_apps.app_store_id\" must be a string, found number", time.Now().Add(24 * time.Hour), map[string]uint{}},
-		{"testdata/gitops/team_vpp_empty_adamid.yml", "software app store id required", time.Now().Add(24 * time.Hour), map[string]uint{}},
-		{"testdata/gitops/team_vpp_valid_app_labels_exclude_any.yml", "", time.Now().Add(24 * time.Hour), map[string]uint{"label 1": 1, "label 2": 2}},
-		{"testdata/gitops/team_vpp_valid_app_labels_include_any.yml", "", time.Now().Add(24 * time.Hour), map[string]uint{"label 1": 1, "label 2": 2}},
-		{"testdata/gitops/team_vpp_invalid_app_labels_exclude_any.yml", "Please create the missing labels, or update your settings to not refer to these labels.", time.Now().Add(24 * time.Hour), map[string]uint{"label 1": 1, "label 2": 2}},
-		{"testdata/gitops/team_vpp_invalid_app_labels_include_any.yml", "Please create the missing labels, or update your settings to not refer to these labels.", time.Now().Add(24 * time.Hour), map[string]uint{"label 1": 1, "label 2": 2}},
-		{"testdata/gitops/team_vpp_invalid_app_labels_both.yml", `only one of "labels_exclude_any" or "labels_include_any" can be specified for app store app`, time.Now().Add(24 * time.Hour), map[string]uint{}},
-	}
-
-	for _, c := range cases {
-		t.Run(filepath.Base(c.file), func(t *testing.T) {
-			ds, _, _ := setupFullGitOpsPremiumServer(t)
-			token, err := test.CreateVPPTokenEncoded(c.tokenExpiration, "fleet", "ca")
-			require.NoError(t, err)
-
-			ds.SetTeamVPPAppsFunc = func(ctx context.Context, teamID *uint, adamIDs []fleet.VPPAppTeam) error {
-				return nil
-			}
-			ds.BatchInsertVPPAppsFunc = func(ctx context.Context, apps []*fleet.VPPApp) error {
-				return nil
-			}
-			ds.GetVPPAppsFunc = func(ctx context.Context, teamID *uint) ([]fleet.VPPAppResponse, error) {
-				return []fleet.VPPAppResponse{}, nil
-			}
-
-			ds.GetVPPTokenByTeamIDFunc = func(ctx context.Context, teamID *uint) (*fleet.VPPTokenDB, error) {
-				return &fleet.VPPTokenDB{
-					ID:        1,
-					OrgName:   "Fleet",
-					Location:  "Earth",
-					RenewDate: c.tokenExpiration,
-					Token:     string(token),
-					Teams:     nil,
-				}, nil
-			}
-
-			ds.GetLabelSpecsFunc = func(ctx context.Context) ([]*fleet.LabelSpec, error) {
-				return []*fleet.LabelSpec{
-					{
-						Name:                "label 1",
-						Description:         "A global label",
-						LabelMembershipType: fleet.LabelMembershipTypeManual,
-						Hosts:               []string{"host2", "host3"},
-					},
-					{
-						Name:                "label 2",
-						Description:         "Another label",
-						LabelMembershipType: fleet.LabelMembershipTypeDynamic,
-						Query:               "SELECT 1 from osquery_info",
-					},
-				}, nil
-			}
-			ds.GetSoftwareCategoryIDsFunc = func(ctx context.Context, names []string) ([]uint, error) {
-				return []uint{}, nil
-			}
-
-			found := make(map[string]uint)
-			ds.LabelIDsByNameFunc = func(ctx context.Context, labels []string) (map[string]uint, error) {
-				for _, l := range labels {
-					if id, ok := c.expectedLabels[l]; ok {
-						found[l] = id
-					}
-				}
-				return found, nil
-			}
-
-			_, err = RunAppNoChecks([]string{"gitops", "-f", c.file})
-
-			if c.wantErr == "" {
-				require.NoError(t, err)
-				if len(c.expectedLabels) > 0 {
-					require.True(t, ds.LabelIDsByNameFuncInvoked)
-				}
-
-				require.Equal(t, c.expectedLabels, found)
-			} else {
-				require.ErrorContains(t, err, c.wantErr)
-			}
-		})
-	}
-}
-
-// TestGitOpsTeamVPPAndApp tests the flow where a new team is created with VPP apps.
-// GitOps must first create the team, then assign VPP token to it, and only then add VPP apps.
-func TestGitOpsTeamVPPAndApp(t *testing.T) {
-	startAndServeVPPServer(t)
-	ds, _, _ := setupFullGitOpsPremiumServer(t)
-	renewDate := time.Now().Add(24 * time.Hour)
-	token, err := test.CreateVPPTokenEncoded(renewDate, "fleet", "ca")
-	require.NoError(t, err)
-
-	ds.GetVPPAppsFunc = func(ctx context.Context, teamID *uint) ([]fleet.VPPAppResponse, error) {
-		return []fleet.VPPAppResponse{}, nil
-	}
-	ds.GetABMTokenCountFunc = func(ctx context.Context) (int, error) {
-		return 0, nil
-	}
-
-	// The following mocks are key to this test.
-	vppToken := &fleet.VPPTokenDB{
-		ID:        1,
-		OrgName:   "Fleet",
-		Location:  "Earth",
-		RenewDate: renewDate,
-		Token:     string(token),
-		Teams:     nil,
-	}
-	tokensByTeams := make(map[uint]*fleet.VPPTokenDB)
-	ds.UpdateVPPTokenTeamsFunc = func(ctx context.Context, id uint, teams []uint) (*fleet.VPPTokenDB, error) {
-		for _, teamID := range teams {
-			tokensByTeams[teamID] = vppToken
-		}
-		return vppToken, nil
-	}
-	ds.ListVPPTokensFunc = func(ctx context.Context) ([]*fleet.VPPTokenDB, error) {
-		return []*fleet.VPPTokenDB{vppToken}, nil
-	}
-	ds.GetVPPTokenByTeamIDFunc = func(ctx context.Context, teamID *uint) (*fleet.VPPTokenDB, error) {
-		if teamID == nil {
-			return vppToken, nil
-		}
-		token, ok := tokensByTeams[*teamID]
-		if !ok {
-			return nil, sql.ErrNoRows
-		}
-		return token, nil
-	}
-	ds.GetSoftwareCategoryIDsFunc = func(ctx context.Context, names []string) ([]uint, error) {
-		return []uint{}, nil
-	}
-
-	buf, err := RunAppNoChecks([]string{"gitops", "-f", "testdata/gitops/global_config_vpp.yml", "-f",
-		"testdata/gitops/team_vpp_valid_app.yml"})
-	require.NoError(t, err)
-	assert.True(t, ds.UpdateVPPTokenTeamsFuncInvoked)
-	assert.True(t, ds.GetVPPTokenByTeamIDFuncInvoked)
-	assert.True(t, ds.SetTeamVPPAppsFuncInvoked)
-	assert.Contains(t, buf.String(), fmt.Sprintf(reapplyingTeamForVPPAppsMsg, teamName))
-}
-
 func TestGitOpsCustomSettings(t *testing.T) {
 	cases := []struct {
 		file    string
@@ -2887,7 +2316,7 @@ func TestGitOpsCustomSettings(t *testing.T) {
 	}
 	for _, c := range cases {
 		t.Run(filepath.Base(c.file), func(t *testing.T) {
-			ds, appCfgPtr, _ := setupFullGitOpsPremiumServer(t)
+			ds, appCfgPtr, _ := testing_utils.SetupFullGitOpsPremiumServer(t)
 			(*appCfgPtr).MDM.EnabledAndConfigured = true
 			(*appCfgPtr).MDM.WindowsEnabledAndConfigured = true
 			ds.GetLabelSpecsFunc = func(ctx context.Context) ([]*fleet.LabelSpec, error) {
@@ -2945,417 +2374,6 @@ func TestGitOpsCustomSettings(t *testing.T) {
 			}
 		})
 	}
-}
-
-type appleVPPConfigSrvConf struct {
-	Assets        []vpp.Asset
-	SerialNumbers []string
-}
-
-func startVPPApplyServer(t *testing.T, config *appleVPPConfigSrvConf) {
-	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if strings.Contains(r.URL.Path, "associate") {
-			var associations vpp.AssociateAssetsRequest
-
-			decoder := json.NewDecoder(r.Body)
-			if err := decoder.Decode(&associations); err != nil {
-				http.Error(w, "invalid request", http.StatusBadRequest)
-				return
-			}
-
-			if len(associations.Assets) == 0 {
-				w.Header().Set("Content-Type", "application/json")
-				w.WriteHeader(http.StatusBadRequest)
-				res := vpp.ErrorResponse{
-					ErrorNumber:  9718,
-					ErrorMessage: "This request doesn't contain an asset, which is a required argument. Change the request to provide an asset.",
-				}
-				if err := json.NewEncoder(w).Encode(res); err != nil {
-					panic(err)
-				}
-				return
-			}
-
-			if len(associations.SerialNumbers) == 0 {
-				w.Header().Set("Content-Type", "application/json")
-				w.WriteHeader(http.StatusBadRequest)
-				res := vpp.ErrorResponse{
-					ErrorNumber:  9719,
-					ErrorMessage: "Either clientUserIds or serialNumbers are required arguments. Change the request to provide assignable users and devices.",
-				}
-				if err := json.NewEncoder(w).Encode(res); err != nil {
-					panic(err)
-				}
-				return
-			}
-
-			var badAssets []vpp.Asset
-			for _, reqAsset := range associations.Assets {
-				var found bool
-				for _, goodAsset := range config.Assets {
-					if reqAsset == goodAsset {
-						found = true
-					}
-				}
-				if !found {
-					badAssets = append(badAssets, reqAsset)
-				}
-			}
-
-			var badSerials []string
-			for _, reqSerial := range associations.SerialNumbers {
-				var found bool
-				for _, goodSerial := range config.SerialNumbers {
-					if reqSerial == goodSerial {
-						found = true
-					}
-				}
-				if !found {
-					badSerials = append(badSerials, reqSerial)
-				}
-			}
-
-			if len(badAssets) != 0 || len(badSerials) != 0 {
-				errMsg := "error associating assets."
-				if len(badAssets) > 0 {
-					var badAdamIds []string
-					for _, asset := range badAssets {
-						badAdamIds = append(badAdamIds, asset.AdamID)
-					}
-					errMsg += fmt.Sprintf(" assets don't exist on account: %s.", strings.Join(badAdamIds, ", "))
-				}
-				if len(badSerials) > 0 {
-					errMsg += fmt.Sprintf(" bad serials: %s.", strings.Join(badSerials, ", "))
-				}
-				res := vpp.ErrorResponse{
-					ErrorInfo: vpp.ResponseErrorInfo{
-						Assets:        badAssets,
-						ClientUserIds: []string{"something"},
-						SerialNumbers: badSerials,
-					},
-					// Not sure what error should be returned on each
-					// error type
-					ErrorNumber:  1,
-					ErrorMessage: errMsg,
-				}
-				w.Header().Set("Content-Type", "application/json")
-				w.WriteHeader(http.StatusBadRequest)
-				if err := json.NewEncoder(w).Encode(res); err != nil {
-					panic(err)
-				}
-			}
-			return
-		}
-
-		if strings.Contains(r.URL.Path, "assets") {
-			// Then we're responding to GetAssets
-			w.Header().Set("Content-Type", "application/json")
-			encoder := json.NewEncoder(w)
-			err := encoder.Encode(map[string][]vpp.Asset{"assets": config.Assets})
-			if err != nil {
-				panic(err)
-			}
-			return
-		}
-
-		resp := []byte(`{"locationName": "Fleet Location One"}`)
-		if strings.Contains(r.URL.RawQuery, "invalidToken") {
-			// This replicates the response sent back from Apple's VPP endpoints when an invalid
-			// token is passed. For more details see:
-			// https://developer.apple.com/documentation/devicemanagement/app_and_book_management/app_and_book_management_legacy/interpreting_error_codes
-			// https://developer.apple.com/documentation/devicemanagement/client_config
-			// https://developer.apple.com/documentation/devicemanagement/errorresponse
-			// Note that the Apple server returns 200 in this case.
-			resp = []byte(`{"errorNumber": 9622,"errorMessage": "Invalid authentication token"}`)
-		}
-
-		if strings.Contains(r.URL.RawQuery, "serverError") {
-			resp = []byte(`{"errorNumber": 9603,"errorMessage": "Internal server error"}`)
-			w.WriteHeader(http.StatusInternalServerError)
-		}
-
-		_, _ = w.Write(resp)
-	}))
-
-	t.Setenv("FLEET_DEV_VPP_URL", srv.URL)
-	t.Cleanup(srv.Close)
-}
-
-func startAndServeVPPServer(t *testing.T) {
-	config := &appleVPPConfigSrvConf{
-		Assets: []vpp.Asset{
-			{
-				AdamID:         "1",
-				PricingParam:   "STDQ",
-				AvailableCount: 12,
-			},
-			{
-				AdamID:         "2",
-				PricingParam:   "STDQ",
-				AvailableCount: 3,
-			},
-		},
-		SerialNumbers: []string{"123", "456"},
-	}
-
-	startVPPApplyServer(t, config)
-
-	appleITunesSrv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		// a map of apps we can respond with
-		db := map[string]string{
-			// macos app
-			"1": `{"bundleId": "a-1", "artworkUrl512": "https://example.com/images/1", "version": "1.0.0", "trackName": "App 1", "TrackID": 1}`,
-			// macos, ios, ipados app
-			"2": `{"bundleId": "b-2", "artworkUrl512": "https://example.com/images/2", "version": "2.0.0", "trackName": "App 2", "TrackID": 2,
-				"supportedDevices": ["MacDesktop-MacDesktop", "iPhone5s-iPhone5s", "iPadAir-iPadAir"] }`,
-			// ipados app
-			"3": `{"bundleId": "c-3", "artworkUrl512": "https://example.com/images/3", "version": "3.0.0", "trackName": "App 3", "TrackID": 3,
-				"supportedDevices": ["iPadAir-iPadAir"] }`,
-		}
-
-		adamIDString := r.URL.Query().Get("id")
-		adamIDs := strings.Split(adamIDString, ",")
-
-		var objs []string
-		for _, a := range adamIDs {
-			objs = append(objs, db[a])
-		}
-
-		_, _ = w.Write([]byte(fmt.Sprintf(`{"results": [%s]}`, strings.Join(objs, ","))))
-	}))
-	t.Setenv("FLEET_DEV_ITUNES_URL", appleITunesSrv.URL)
-}
-
-func setupFullGitOpsPremiumServer(t *testing.T) (*mock.Store, **fleet.AppConfig, map[string]**fleet.Team) {
-	testCert, testKey, err := apple_mdm.NewSCEPCACertKey()
-	require.NoError(t, err)
-	testCertPEM := tokenpki.PEMCertificate(testCert.Raw)
-	testKeyPEM := tokenpki.PEMRSAPrivateKey(testKey)
-	fleetCfg := config.TestConfig()
-	config.SetTestMDMConfig(t, &fleetCfg, testCertPEM, testKeyPEM, "../../../server/service/testdata")
-
-	license := &fleet.LicenseInfo{Tier: fleet.TierPremium, Expiration: time.Now().Add(24 * time.Hour)}
-	_, ds := testingutils.RunServerWithMockedDS(
-		t, &service.TestServerOpts{
-			MDMStorage:       new(mdmmock.MDMAppleStore),
-			MDMPusher:        mockPusher{},
-			FleetConfig:      &fleetCfg,
-			License:          license,
-			NoCacheDatastore: true,
-			KeyValueStore:    newMemKeyValueStore(),
-		},
-	)
-
-	// Mock appConfig
-	savedAppConfig := &fleet.AppConfig{
-		MDM: fleet.MDM{
-			EnabledAndConfigured: true,
-		},
-	}
-	addLabelMocks(ds)
-
-	ds.AppConfigFunc = func(ctx context.Context) (*fleet.AppConfig, error) {
-		appConfigCopy := *savedAppConfig
-		return &appConfigCopy, nil
-	}
-	ds.SaveAppConfigFunc = func(ctx context.Context, config *fleet.AppConfig) error {
-		appConfigCopy := *config
-		savedAppConfig = &appConfigCopy
-		return nil
-	}
-	ds.SetTeamVPPAppsFunc = func(ctx context.Context, teamID *uint, adamIDs []fleet.VPPAppTeam) error {
-		return nil
-	}
-	ds.BatchInsertVPPAppsFunc = func(ctx context.Context, apps []*fleet.VPPApp) error {
-		return nil
-	}
-
-	savedTeams := map[string]**fleet.Team{}
-
-	ds.ApplyEnrollSecretsFunc = func(ctx context.Context, teamID *uint, secrets []*fleet.EnrollSecret) error {
-		return nil
-	}
-	ds.ApplyPolicySpecsFunc = func(ctx context.Context, authorID uint, specs []*fleet.PolicySpec) error {
-		return nil
-	}
-	ds.ApplyQueriesFunc = func(
-		ctx context.Context, authorID uint, queries []*fleet.Query, queriesToDiscardResults map[uint]struct{},
-	) error {
-		return nil
-	}
-	ds.BatchSetMDMProfilesFunc = func(
-		ctx context.Context, tmID *uint, macProfiles []*fleet.MDMAppleConfigProfile, winProfiles []*fleet.MDMWindowsConfigProfile,
-		macDecls []*fleet.MDMAppleDeclaration, vars []fleet.MDMProfileIdentifierFleetVariables,
-	) (updates fleet.MDMProfilesUpdates, err error) {
-		return fleet.MDMProfilesUpdates{}, nil
-	}
-	ds.BatchSetScriptsFunc = func(ctx context.Context, tmID *uint, scripts []*fleet.Script) ([]fleet.ScriptResponse, error) {
-		return []fleet.ScriptResponse{}, nil
-	}
-	ds.BulkSetPendingMDMHostProfilesFunc = func(
-		ctx context.Context, hostIDs []uint, teamIDs []uint, profileUUIDs []string, hostUUIDs []string,
-	) (updates fleet.MDMProfilesUpdates, err error) {
-		return fleet.MDMProfilesUpdates{}, nil
-	}
-	ds.DeleteMDMAppleDeclarationByNameFunc = func(ctx context.Context, teamID *uint, name string) error {
-		return nil
-	}
-	ds.GetMDMAppleBootstrapPackageMetaFunc = func(ctx context.Context, teamID uint) (*fleet.MDMAppleBootstrapPackage, error) {
-		return &fleet.MDMAppleBootstrapPackage{}, nil
-	}
-	ds.DeleteMDMAppleBootstrapPackageFunc = func(ctx context.Context, teamID uint) error {
-		return nil
-	}
-	ds.GetMDMAppleSetupAssistantFunc = func(ctx context.Context, teamID *uint) (*fleet.MDMAppleSetupAssistant, error) {
-		return nil, nil
-	}
-	ds.DeleteMDMAppleSetupAssistantFunc = func(ctx context.Context, teamID *uint) error {
-		return nil
-	}
-	ds.IsEnrollSecretAvailableFunc = func(ctx context.Context, secret string, isNew bool, teamID *uint) (bool, error) {
-		return true, nil
-	}
-	ds.LabelIDsByNameFunc = func(ctx context.Context, labels []string) (map[string]uint, error) {
-		require.ElementsMatch(t, labels, []string{fleet.BuiltinLabelMacOS14Plus})
-		return map[string]uint{fleet.BuiltinLabelMacOS14Plus: 1}, nil
-	}
-	ds.ListGlobalPoliciesFunc = func(ctx context.Context, opts fleet.ListOptions) ([]*fleet.Policy, error) { return nil, nil }
-	ds.ListTeamPoliciesFunc = func(
-		ctx context.Context, teamID uint, opts fleet.ListOptions, iopts fleet.ListOptions,
-	) (teamPolicies []*fleet.Policy, inheritedPolicies []*fleet.Policy, err error) {
-		return nil, nil, nil
-	}
-	ds.ListTeamsFunc = func(ctx context.Context, filter fleet.TeamFilter, opt fleet.ListOptions) ([]*fleet.Team, error) {
-		if savedTeams != nil {
-			var result []*fleet.Team
-			for _, t := range savedTeams {
-				result = append(result, *t)
-			}
-			return result, nil
-		}
-		return nil, nil
-	}
-	ds.TeamsSummaryFunc = func(ctx context.Context) ([]*fleet.TeamSummary, error) {
-		summary := make([]*fleet.TeamSummary, 0, len(savedTeams))
-		for _, team := range savedTeams {
-			summary = append(summary, &fleet.TeamSummary{
-				ID:          (*team).ID,
-				Name:        (*team).Name,
-				Description: (*team).Description,
-			})
-		}
-		return summary, nil
-	}
-	ds.ListQueriesFunc = func(ctx context.Context, opts fleet.ListQueryOptions) ([]*fleet.Query, int, *fleet.PaginationMetadata, error) {
-		return nil, 0, nil, nil
-	}
-	ds.NewActivityFunc = func(
-		ctx context.Context, user *fleet.User, activity fleet.ActivityDetails, details []byte, createdAt time.Time,
-	) error {
-		return nil
-	}
-	ds.NewMDMAppleConfigProfileFunc = func(ctx context.Context, p fleet.MDMAppleConfigProfile, vars []string) (*fleet.MDMAppleConfigProfile, error) {
-		return nil, nil
-	}
-	ds.NewJobFunc = func(ctx context.Context, job *fleet.Job) (*fleet.Job, error) {
-		job.ID = 1
-		return job, nil
-	}
-	ds.NewTeamFunc = func(ctx context.Context, team *fleet.Team) (*fleet.Team, error) {
-		team.ID = uint(len(savedTeams) + 1) //nolint:gosec // dismiss G115
-		savedTeams[team.Name] = &team
-		return team, nil
-	}
-	ds.QueryByNameFunc = func(ctx context.Context, teamID *uint, name string) (*fleet.Query, error) {
-		return nil, &notFoundError{}
-	}
-	ds.TeamFunc = func(ctx context.Context, tid uint) (*fleet.Team, error) {
-		for _, tm := range savedTeams {
-			if (*tm).ID == tid {
-				return *tm, nil
-			}
-		}
-		return nil, &notFoundError{}
-	}
-	ds.TeamByNameFunc = func(ctx context.Context, name string) (*fleet.Team, error) {
-		for _, tm := range savedTeams {
-			if (*tm).Name == name {
-				return *tm, nil
-			}
-		}
-		return nil, &notFoundError{}
-	}
-	ds.TeamByFilenameFunc = func(ctx context.Context, filename string) (*fleet.Team, error) {
-		for _, tm := range savedTeams {
-			if (*tm).Filename != nil && *(*tm).Filename == filename {
-				return *tm, nil
-			}
-		}
-		return nil, &notFoundError{}
-	}
-	ds.SaveTeamFunc = func(ctx context.Context, team *fleet.Team) (*fleet.Team, error) {
-		savedTeams[team.Name] = &team
-		return team, nil
-	}
-	ds.SetOrUpdateMDMAppleDeclarationFunc = func(ctx context.Context, declaration *fleet.MDMAppleDeclaration) (
-		*fleet.MDMAppleDeclaration, error,
-	) {
-		declaration.DeclarationUUID = uuid.NewString()
-		return declaration, nil
-	}
-	ds.BatchSetSoftwareInstallersFunc = func(ctx context.Context, teamID *uint, installers []*fleet.UploadSoftwareInstallerPayload) error {
-		return nil
-	}
-	ds.GetSoftwareInstallersFunc = func(ctx context.Context, tmID uint) ([]fleet.SoftwarePackageResponse, error) {
-		return nil, nil
-	}
-
-	ds.InsertVPPTokenFunc = func(ctx context.Context, tok *fleet.VPPTokenData) (*fleet.VPPTokenDB, error) {
-		return &fleet.VPPTokenDB{}, nil
-	}
-	ds.GetVPPTokenFunc = func(ctx context.Context, tokenID uint) (*fleet.VPPTokenDB, error) {
-		return &fleet.VPPTokenDB{}, err
-	}
-	ds.GetVPPTokenByTeamIDFunc = func(ctx context.Context, teamID *uint) (*fleet.VPPTokenDB, error) {
-		return &fleet.VPPTokenDB{}, nil
-	}
-	ds.ListVPPTokensFunc = func(ctx context.Context) ([]*fleet.VPPTokenDB, error) {
-		return nil, nil
-	}
-	ds.UpdateVPPTokenTeamsFunc = func(ctx context.Context, id uint, teams []uint) (*fleet.VPPTokenDB, error) {
-		return &fleet.VPPTokenDB{}, nil
-	}
-	ds.ListABMTokensFunc = func(ctx context.Context) ([]*fleet.ABMToken, error) {
-		return []*fleet.ABMToken{{OrganizationName: "Fleet Device Management Inc."}}, nil
-	}
-	ds.ListSoftwareTitlesFunc = func(ctx context.Context, opt fleet.SoftwareTitleListOptions, tmFilter fleet.TeamFilter) ([]fleet.SoftwareTitleListResult, int, *fleet.PaginationMetadata, error) {
-		return nil, 0, nil, nil
-	}
-	ds.SaveABMTokenFunc = func(ctx context.Context, tok *fleet.ABMToken) error {
-		return nil
-	}
-	ds.ListVPPTokensFunc = func(ctx context.Context) ([]*fleet.VPPTokenDB, error) {
-		return []*fleet.VPPTokenDB{}, nil
-	}
-	ds.ListABMTokensFunc = func(ctx context.Context) ([]*fleet.ABMToken, error) {
-		return []*fleet.ABMToken{}, nil
-	}
-	ds.DeleteSetupExperienceScriptFunc = func(ctx context.Context, teamID *uint) error {
-		return nil
-	}
-	ds.SetSetupExperienceScriptFunc = func(ctx context.Context, script *fleet.Script) error {
-		return nil
-	}
-	ds.ExpandEmbeddedSecretsAndUpdatedAtFunc = func(ctx context.Context, document string) (string, *time.Time, error) {
-		return document, nil, nil
-	}
-
-	t.Setenv("FLEET_SERVER_URL", fleetServerURL)
-	t.Setenv("ORG_NAME", orgName)
-	t.Setenv("TEST_TEAM_NAME", teamName)
-	t.Setenv("APPLE_BM_DEFAULT_TEAM", teamName)
-
-	return ds, &savedAppConfig, savedTeams
 }
 
 func TestGitOpsABM(t *testing.T) {
@@ -3662,7 +2680,7 @@ software:
 
 	for _, tt := range cases {
 		t.Run(tt.name, func(t *testing.T) {
-			ds, savedAppConfigPtr, savedTeams := setupFullGitOpsPremiumServer(t)
+			ds, savedAppConfigPtr, savedTeams := testing_utils.SetupFullGitOpsPremiumServer(t)
 
 			ds.ListABMTokensFunc = func(ctx context.Context) ([]*fleet.ABMToken, error) {
 				if len(tt.tokens) > 0 {
@@ -3672,332 +2690,6 @@ software:
 			}
 			ds.GetABMTokenCountFunc = func(ctx context.Context) (int, error) {
 				return len(tt.tokens), nil
-			}
-
-			ds.TeamsSummaryFunc = func(ctx context.Context) ([]*fleet.TeamSummary, error) {
-				var res []*fleet.TeamSummary
-				for _, tm := range savedTeams {
-					res = append(res, &fleet.TeamSummary{Name: (*tm).Name, ID: (*tm).ID})
-				}
-				return res, nil
-			}
-
-			ds.SaveABMTokenFunc = func(ctx context.Context, tok *fleet.ABMToken) error {
-				return nil
-			}
-
-			args := []string{"gitops"}
-			for _, cfg := range tt.cfgs {
-				if cfg != "" {
-					tmpFile, err := os.CreateTemp(t.TempDir(), "*.yml")
-					require.NoError(t, err)
-					_, err = tmpFile.WriteString(cfg)
-					require.NoError(t, err)
-					args = append(args, "-f", tmpFile.Name())
-				}
-			}
-
-			// Dry run
-			out, err := RunAppNoChecks(append(args, "--dry-run"))
-			tt.dryRunAssertion(t, *savedAppConfigPtr, ds, out.String(), err)
-			if t.Failed() {
-				t.FailNow()
-			}
-
-			// Real run
-			out, err = RunAppNoChecks(args)
-			tt.realRunAssertion(t, *savedAppConfigPtr, ds, out.String(), err)
-
-			// Second real run, now that all the teams are saved
-			out, err = RunAppNoChecks(args)
-			tt.realRunAssertion(t, *savedAppConfigPtr, ds, out.String(), err)
-		})
-	}
-}
-
-func TestGitOpsVPP(t *testing.T) {
-	global := func(mdm string) string {
-		return fmt.Sprintf(`
-controls:
-queries:
-policies:
-agent_options:
-software:
-org_settings:
-  server_settings:
-    server_url: "https://foo.example.com"
-  org_info:
-    org_name: GitOps Test
-  secrets:
-    - secret: "global"
-  mdm:
-    %s
- `, mdm)
-	}
-
-	team := func(name string) string {
-		return fmt.Sprintf(`
-name: %s
-team_settings:
-  secrets:
-    - secret: "%s-secret"
-agent_options:
-controls:
-policies:
-queries:
-software:
-`, name, name)
-	}
-
-	workstations := team("💻 Workstations")
-	iosTeam := team("📱🏢 Company-owned iPhones")
-	ipadTeam := team("🔳🏢 Company-owned iPads")
-
-	cases := []struct {
-		name             string
-		cfgs             []string
-		dryRunAssertion  func(t *testing.T, appCfg *fleet.AppConfig, ds fleet.Datastore, out string, err error)
-		realRunAssertion func(t *testing.T, appCfg *fleet.AppConfig, ds fleet.Datastore, out string, err error)
-	}{
-		{
-			name: "new key all valid",
-			cfgs: []string{
-				global(`
-                                  volume_purchasing_program:
-                                    - location: Fleet Device Management Inc.
-                                      teams:
-                                        - "💻 Workstations"
-                                        - "📱🏢 Company-owned iPhones"
-                                        - "🔳🏢 Company-owned iPads"`),
-				workstations,
-				iosTeam,
-				ipadTeam,
-			},
-			dryRunAssertion: func(t *testing.T, appCfg *fleet.AppConfig, ds fleet.Datastore, out string, err error) {
-				assert.NoError(t, err)
-				assert.Empty(t, appCfg.MDM.VolumePurchasingProgram.Value)
-				assert.Contains(t, out, "[!] gitops dry run succeeded")
-			},
-			realRunAssertion: func(t *testing.T, appCfg *fleet.AppConfig, ds fleet.Datastore, out string, err error) {
-				assert.NoError(t, err)
-				assert.ElementsMatch(
-					t,
-					appCfg.MDM.VolumePurchasingProgram.Value,
-					[]fleet.MDMAppleVolumePurchasingProgramInfo{
-						{
-							Location: "Fleet Device Management Inc.",
-							Teams: []string{
-								"💻 Workstations",
-								"📱🏢 Company-owned iPhones",
-								"🔳🏢 Company-owned iPads",
-							},
-						},
-					},
-				)
-				assert.Contains(t, out, "[!] gitops succeeded")
-			},
-		},
-		{
-			name: "new key multiple elements",
-			cfgs: []string{
-				global(`
-                                  volume_purchasing_program:
-                                    - location: Acme Inc.
-                                      teams:
-                                        - "💻 Workstations"
-                                    - location: Fleet Device Management Inc.
-                                      teams:
-                                        - "📱🏢 Company-owned iPhones"
-                                        - "🔳🏢 Company-owned iPads"`),
-				workstations,
-				iosTeam,
-				ipadTeam,
-			},
-			dryRunAssertion: func(t *testing.T, appCfg *fleet.AppConfig, ds fleet.Datastore, out string, err error) {
-				assert.NoError(t, err)
-				assert.Empty(t, appCfg.MDM.VolumePurchasingProgram.Value)
-				assert.Contains(t, out, "[!] gitops dry run succeeded")
-			},
-			realRunAssertion: func(t *testing.T, appCfg *fleet.AppConfig, ds fleet.Datastore, out string, err error) {
-				assert.NoError(t, err)
-				assert.ElementsMatch(
-					t,
-					appCfg.MDM.VolumePurchasingProgram.Value,
-					[]fleet.MDMAppleVolumePurchasingProgramInfo{
-						{
-							Location: "Acme Inc.",
-							Teams: []string{
-								"💻 Workstations",
-							},
-						},
-						{
-							Location: "Fleet Device Management Inc.",
-							Teams: []string{
-								"📱🏢 Company-owned iPhones",
-								"🔳🏢 Company-owned iPads",
-							},
-						},
-					},
-				)
-				assert.Contains(t, out, "[!] gitops succeeded")
-			},
-		},
-		{
-			name: "using an undefined team errors",
-			cfgs: []string{
-				global(`
-                                  volume_purchasing_program:
-                                    - location: Fleet Device Management Inc.
-                                      teams:
-                                        - "💻 Workstations"
-                                        - "📱🏢 Company-owned iPhones"
-                                        - "🔳🏢 Company-owned iPads"`),
-				workstations,
-				ipadTeam,
-			},
-			dryRunAssertion: func(t *testing.T, appCfg *fleet.AppConfig, ds fleet.Datastore, out string, err error) {
-				assert.ErrorContains(t, err, "volume_purchasing_program team 📱🏢 Company-owned iPhones not found in team configs")
-			},
-			realRunAssertion: func(t *testing.T, appCfg *fleet.AppConfig, ds fleet.Datastore, out string, err error) {
-				assert.ErrorContains(t, err, "volume_purchasing_program team 📱🏢 Company-owned iPhones not found in team configs")
-			},
-		},
-		{
-			name: "no team is supported",
-			cfgs: []string{
-				global(`
-                                  volume_purchasing_program:
-                                    - location: Fleet Device Management Inc.
-                                      teams:
-                                        - "💻 Workstations"
-                                        - "📱🏢 Company-owned iPhones"
-                                        - "No team"`),
-				workstations,
-				iosTeam,
-			},
-			dryRunAssertion: func(t *testing.T, appCfg *fleet.AppConfig, ds fleet.Datastore, out string, err error) {
-				assert.NoError(t, err)
-				assert.Empty(t, appCfg.MDM.VolumePurchasingProgram.Value)
-				assert.Contains(t, out, "[!] gitops dry run succeeded")
-			},
-			realRunAssertion: func(t *testing.T, appCfg *fleet.AppConfig, ds fleet.Datastore, out string, err error) {
-				assert.NoError(t, err)
-				assert.ElementsMatch(
-					t,
-					appCfg.MDM.VolumePurchasingProgram.Value,
-					[]fleet.MDMAppleVolumePurchasingProgramInfo{
-						{
-							Location: "Fleet Device Management Inc.",
-							Teams: []string{
-								"💻 Workstations",
-								"📱🏢 Company-owned iPhones",
-								"No team",
-							},
-						},
-					},
-				)
-				assert.Contains(t, out, "[!] gitops succeeded")
-			},
-		},
-		{
-			name: "all teams is supported",
-			cfgs: []string{
-				global(`
-                        volume_purchasing_program:
-                          - location: Fleet Device Management Inc.
-                            teams:
-                              - "All teams"`),
-				workstations,
-				iosTeam,
-			},
-			dryRunAssertion: func(t *testing.T, appCfg *fleet.AppConfig, ds fleet.Datastore, out string, err error) {
-				assert.NoError(t, err)
-				assert.Empty(t, appCfg.MDM.VolumePurchasingProgram.Value)
-				assert.Contains(t, out, "[!] gitops dry run succeeded")
-			},
-			realRunAssertion: func(t *testing.T, appCfg *fleet.AppConfig, ds fleet.Datastore, out string, err error) {
-				assert.NoError(t, err)
-				assert.ElementsMatch(
-					t,
-					appCfg.MDM.VolumePurchasingProgram.Value,
-					[]fleet.MDMAppleVolumePurchasingProgramInfo{
-						{
-							Location: "Fleet Device Management Inc.",
-							Teams: []string{
-								"All teams",
-							},
-						},
-					},
-				)
-				assert.Contains(t, out, "[!] gitops succeeded")
-			},
-		},
-		{
-			name: "not provided teams defaults to no team",
-			cfgs: []string{
-				global(`
-                                  volume_purchasing_program:
-                                    - location: Fleet Device Management Inc.
-                                      teams:`),
-				workstations,
-				ipadTeam,
-			},
-			dryRunAssertion: func(t *testing.T, appCfg *fleet.AppConfig, ds fleet.Datastore, out string, err error) {
-				assert.NoError(t, err)
-				assert.Empty(t, appCfg.MDM.VolumePurchasingProgram.Value)
-				assert.Contains(t, out, "[!] gitops dry run succeeded")
-			},
-			realRunAssertion: func(t *testing.T, appCfg *fleet.AppConfig, ds fleet.Datastore, out string, err error) {
-				assert.NoError(t, err)
-				assert.ElementsMatch(
-					t,
-					appCfg.MDM.VolumePurchasingProgram.Value,
-					[]fleet.MDMAppleVolumePurchasingProgramInfo{
-						{
-							Location: "Fleet Device Management Inc.",
-							Teams:    nil,
-						},
-					},
-				)
-				assert.Contains(t, out, "[!] gitops succeeded")
-			},
-		},
-		{
-			name: "non existent location fails",
-			cfgs: []string{
-				global(`
-                                  volume_purchasing_program:
-                                    - location: Does not exist
-                                      teams:`),
-				workstations,
-				ipadTeam,
-			},
-			dryRunAssertion: func(t *testing.T, appCfg *fleet.AppConfig, ds fleet.Datastore, out string, err error) {
-				assert.ErrorContains(t, err, "token with location Does not exist doesn't exist")
-				assert.Empty(t, appCfg.MDM.VolumePurchasingProgram.Value)
-				assert.NotContains(t, out, "[!] gitops dry run succeeded")
-			},
-			realRunAssertion: func(t *testing.T, appCfg *fleet.AppConfig, ds fleet.Datastore, out string, err error) {
-				assert.ErrorContains(t, err, "token with location Does not exist doesn't exist")
-				assert.Empty(t, appCfg.MDM.VolumePurchasingProgram.Value)
-				assert.NotContains(t, out, "[!] gitops dry run succeeded")
-			},
-		},
-	}
-
-	for _, tt := range cases {
-		t.Run(tt.name, func(t *testing.T) {
-			ds, savedAppConfigPtr, savedTeams := setupFullGitOpsPremiumServer(t)
-
-			ds.ListVPPTokensFunc = func(ctx context.Context) ([]*fleet.VPPTokenDB, error) {
-				return []*fleet.VPPTokenDB{{Location: "Fleet Device Management Inc."}, {Location: "Acme Inc."}}, nil
-			}
-
-			ds.ListABMTokensFunc = func(ctx context.Context) ([]*fleet.ABMToken, error) {
-				return []*fleet.ABMToken{{OrganizationName: "Fleet Device Management Inc."}, {OrganizationName: "Foo Inc."}}, nil
-			}
-			ds.GetABMTokenCountFunc = func(ctx context.Context) (int, error) {
-				return 1, nil
 			}
 
 			ds.TeamsSummaryFunc = func(ctx context.Context) ([]*fleet.TeamSummary, error) {
@@ -4054,7 +2746,7 @@ func TestGitOpsWindowsMigration(t *testing.T) {
 	}
 	for _, c := range cases {
 		t.Run(filepath.Base(c.file), func(t *testing.T) {
-			setupFullGitOpsPremiumServer(t)
+			testing_utils.SetupFullGitOpsPremiumServer(t)
 
 			_, err := RunAppNoChecks([]string{"gitops", "-f", c.file})
 			if c.wantErr == "" {
@@ -4067,7 +2759,7 @@ func TestGitOpsWindowsMigration(t *testing.T) {
 }
 
 func TestGitOpsGlobalWebhooksDisable(t *testing.T) {
-	_, appConfig, _ := setupFullGitOpsPremiumServer(t)
+	_, appConfig, _ := testing_utils.SetupFullGitOpsPremiumServer(t)
 
 	webhook := &(*appConfig).WebhookSettings
 	webhook.ActivitiesWebhook.Enable = true
@@ -4089,7 +2781,7 @@ func TestGitOpsGlobalWebhooksDisable(t *testing.T) {
 func TestGitOpsTeamWebhooks(t *testing.T) {
 	teamName := "TestTeamWebhooks"
 
-	ds, _, savedTeams := setupFullGitOpsPremiumServer(t)
+	ds, _, savedTeams := testing_utils.SetupFullGitOpsPremiumServer(t)
 
 	// Create a new team.
 	_, err := ds.NewTeam(context.Background(), &fleet.Team{Name: teamName, Config: fleet.TeamConfig{WebhookSettings: fleet.TeamWebhookSettings{
@@ -4118,24 +2810,3 @@ func TestGitOpsTeamWebhooks(t *testing.T) {
 	require.Equal(t, "http://coolwebhook.biz", team.Config.WebhookSettings.HostStatusWebhook.DestinationURL)
 }
 
-type memKeyValueStore struct {
-	m sync.Map
-}
-
-func newMemKeyValueStore() *memKeyValueStore {
-	return &memKeyValueStore{}
-}
-
-func (m *memKeyValueStore) Set(ctx context.Context, key string, value string, expireTime time.Duration) error {
-	m.m.Store(key, value)
-	return nil
-}
-
-func (m *memKeyValueStore) Get(ctx context.Context, key string) (*string, error) {
-	v, ok := m.m.Load(key)
-	if !ok {
-		return nil, nil
-	}
-	vAsString := v.(string)
-	return &vAsString, nil
-}
