+++ cmd/fleetctl/fleetctl/mdm_test.go
@@ -9,10 +9,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/fleetdm/fleet/v4/cmd/fleetctl/fleetctl/testingutils"
+	"github.com/fleetdm/fleet/v4/cmd/fleetctl/fleetctl/testing_utils"
 	"github.com/fleetdm/fleet/v4/server/fleet"
 	"github.com/fleetdm/fleet/v4/server/mdm/nanomdm/mdm"
-	"github.com/fleetdm/fleet/v4/server/mdm/nanomdm/push"
 	"github.com/fleetdm/fleet/v4/server/mock"
 	mdmmock "github.com/fleetdm/fleet/v4/server/mock/mdm"
 	"github.com/fleetdm/fleet/v4/server/ptr"
@@ -21,19 +20,9 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-type mockPusher struct{}
-
 type testhost struct {
 	host    *fleet.Host
 	mdmInfo *fleet.HostMDM
-}
-
-func (mockPusher) Push(ctx context.Context, ids []string) (map[string]*push.Response, error) {
-	m := make(map[string]*push.Response, len(ids))
-	for _, id := range ids {
-		m[id] = &push.Response{Id: id}
-	}
-	return m, nil
 }
 
 func TestMDMRunCommand(t *testing.T) {
@@ -193,9 +182,9 @@ func TestMDMRunCommand(t *testing.T) {
 			enqueuer := new(mdmmock.MDMAppleStore)
 			license := &fleet.LicenseInfo{Tier: lic, Expiration: time.Now().Add(24 * time.Hour)}
 
-			_, ds := testingutils.RunServerWithMockedDS(t, &service.TestServerOpts{
+			_, ds := testing_utils.RunServerWithMockedDS(t, &service.TestServerOpts{
 				MDMStorage:       enqueuer,
-				MDMPusher:        mockPusher{},
+				MDMPusher: testing_utils.MockPusher{},
 				License:          license,
 				NoCacheDatastore: true,
 			})
@@ -1323,9 +1312,9 @@ func setupTestServer(t *testing.T) *mock.Store {
 		return nil
 	}
 
-	_, ds := testingutils.RunServerWithMockedDS(t, &service.TestServerOpts{
+	_, ds := testing_utils.RunServerWithMockedDS(t, &service.TestServerOpts{
 		MDMStorage:       enqueuer,
-		MDMPusher:        mockPusher{},
+		MDMPusher: testing_utils.MockPusher{},
 		License:          license,
 		NoCacheDatastore: true,
 	})
