+++ cmd/fleetctl/fleetctl/trigger_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/fleetdm/fleet/v4/cmd/fleetctl/fleetctl/testingutils"
+	"github.com/fleetdm/fleet/v4/cmd/fleetctl/fleetctl/testing_utils"
 	"github.com/fleetdm/fleet/v4/server/fleet"
 	"github.com/fleetdm/fleet/v4/server/service"
 	"github.com/fleetdm/fleet/v4/server/service/schedule"
@@ -53,7 +53,7 @@ func TestTrigger(t *testing.T) {
 	oldStdout := os.Stdout
 	os.Stdout = w
 
-	_, _ = testingutils.RunServerWithMockedDS(t, &service.TestServerOpts{
+	_, _ = testing_utils.RunServerWithMockedDS(t, &service.TestServerOpts{
 		Logger: kitlog.NewNopLogger(),
 		StartCronSchedules: []service.TestNewScheduleFunc{
 			func(ctx context.Context, ds fleet.Datastore) fleet.NewCronScheduleFunc {
